[LCR-181] refactor: dto 패키지 이동
